Extract render module from main
